changed doule to float in lines 9,20,21,22,23

+++ radialWavefunction2P.cpp
@@ -6,7 +6,7 @@ int main(){
 	float bohr_Radius= 5.29E-11;
 	float r= bohr_Radius*2;
 	int Z,S,Zf;
-	double answer;
+	float answer;
 	std::cout<<"non-valence electrons:";
 	std::cin>>S;
 	std::cout<<std::endl;	std::cout<<"atomic no.:";
@@ -17,10 +17,10 @@ int main(){
 	p1 is 1/root6, p2 is p*Z^3/2 and p3 is e^-p/2
 	*/
 	
-	double p1= 1/(sqrt(6)*2);
-	double rho= Zf*r;
-	double p2=rho*(pow(Zf,(3/2)));
-	double p3=exp(-rho/2);
+	float p1= 1/(sqrt(6)*2);
+	float rho= Zf*r;
+	float p2=rho*(pow(Zf,(3/2)));
+	float p3=exp(-rho/2);
 	
 	answer=p1*p2*p3;
 	std::cout<<"answer:"<<answer<<std::endl;
